requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coverage==5.3.1
Django==3.0.5
djangorestframework==3.12.2
djangotoolbox==1.8.0
djongo==1.3.3
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pymongo==3.11.2
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
pytz==2020.4
requests==2.25.1
sqlparse==0.2.4
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
